🚚 Move: 불필요한 util폴더 한 개 제거 및 파일 이동
